adds retries to cicd tests

+++ .circleci/doTests.sh
@@ -74,16 +74,33 @@ while [ $i -lt $frontendDriverLength ]; do
     fi
     coreFree=$(echo $coreFree | jq .core | tr -d '"')
     someTestsRan=true
-    ./setupAndTestWithFreeCore.sh $coreFree $driverTag
-    if [[ $? -ne 0 ]]
-    then
-        echo "test failed... exiting!"
-        exit 1
+    tries=1
+        while [ $tries -le 3 ]
+        do
+            tries=$(( $tries + 1 ))
+            ./setupAndTestWithFreeCore.sh $coreFree $driverTag
+            if [[ $? -ne 0 ]]
+            then
+                if [[ $tries -le 3 ]]
+                then
+                    rm -rf ../../supertokens-root
+                    rm -rf ../test/server/node_modules/supertokens-node
+                    git checkout HEAD -- ../test/server/package.json
+                    echo "failed test.. retrying!"
+                else
+                    echo "test failed... exiting!"
+                    exit 1
+                fi
+            else
+                rm -rf ../../supertokens-root
+                rm -rf ../test/server/node_modules/supertokens-node
+                git checkout HEAD -- ../test/server/package.json
+                break
+            fi
+        done
     fi
-    rm -rf ../../supertokens-root
-    rm -rf ../test/server/node_modules/supertokens-node
-    git checkout HEAD -- ../test/server/package.json
 done
+
 if [[ $someTestsRan = "true" ]]
 then
     echo "calling /frontend PATCH to make testing passed"
